csgo/syntropy-csgo-bot/requirements.txt
aiohttp==3.6.3
aioify==0.3.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
docker==4.4.1
idna==2.10
idna-ssl==1.1.0
module-wrapper==0.2.4
multidict==4.7.6
poetry-version==0.1.5
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
syntropy-sdk==0.1.2
tomlkit==0.5.11
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.5.1
